Better detect successful activation
When submitting a guest-pass, we check to see if we still have a
free-trail offer thus indicating we are not premium. However, we don't
always wait for a new page to load so we might read stale data. We now
wait until a new page is loaded.

Removed Status enum